Fix #8 - Refactor project detection
This refactors most of the detection and loading of a Trellis project.

The main bug in the original implementation was the use of `os.Chdir` as
the detect function traversed up the directory tree. When no project was
detected, the working directory was still changed to a root directory.
This likely wouldn't have the proper permissions, but even if it did,
the path the project was created at would be entirely wrong.

Now `os.Chdir` only happens when a valid Trellis project is actually
detected. This also only happens now for commands which need to act on a
project and call `LoadProject` explicitly.