Add droplet command
This adds a `droplet` command with one subcommand: `create`

It will create a DigitalOcean droplet (server) via their API and
automatically provision it.

A `DIGITALOCEAN_ACCESS_TOKEN` environment variable is needed to run this
command. If it's not found, a prompt will export it.

This command also requires an SSH key to be added to your DO account.
By default it will look in `~/.ssh/id_rsa.pub` and add it to your
account if it's not already there. A different key path can be specified
with the `--ssh-key` option.

trellis-cli will create a Ubuntu 18.04 x64 server. The region and size
of the droplet is up to the user either via options or prompt
selections.

The name of the droplet will default to the first site name in your
Trellis `wordpress_sites`. This can be changed via a prompt as well.

The droplet will be created with two tags:
* `trellis`
* environment name

Once the server is created, `ansible-galaxy install` will be run to
ensure the Galaxy roles exist. Then the server will be provisioned via
the CLI.