Make autocompletion lazier
These autocomplete functions are called *every* time the binary is run
which was calling `LoadProject` in situations we don't actually want it
loaded.

The only situations we actually want a project loaded is when we're
autocompleting for the commands which is what will happen now since the
checks are *within* the `PredictFunc`.