Unify execCommand + logCmd
Adds a new `execCommandWithOutput` function to combine the previous two usages.
This makes it easier to replace it with a test implementation.

This means all `syscall.Exec` usage can be removed as well now.

syscall usage had a few problems:
* not consistent with other exec usages (via `exec.Command`)
* worked on unix only making this binary much less portable
* much harder to test

This replaces all `syscall.Exec` usage and removes our `CommandExecutor`
abstraction which existed to make testing possible.

A new `CommandExec` function was added for use in these situations where
real Stdout/Stdin was needed (ssh, vault edit so far).

The only difference from the more common `CommandExecWithOutput` is that the
output/input is not set to cli.Ui ones.