Add version update notices
This checks for newer versions of trellis-cli on every run (when
enabled) and displays a notice when there's on available.

This is borrowed almost exactly from [GitHub's CLI](https://github.com/cli/cli).
It's optimized to only check once every 24 hours and the result is
cached locally to avoid HTTP fetches.