Fix #206 - Support virtualenv in paths with spaces
Updates the shebang lines in pip generated bin files to properly handle
paths with spaces.

Pip does not properly handle paths with spaces in them when creating the bin
scripts:

  #!/path with spaces/bin/python

This is an invalid POSIX path so Python can't execute the script.

As a workaround, this function replaces that invalid shebang with the workaround
that Virtualenv uses itself for the pip binary:

  #!/bin/sh
  '''exec' "/path with spaces/bin/python" "$0" "$@"
  ' '''

This shebang fix is only done when the path contain spaces. Otherwise,
nothing is changed.