Fix provision in dev for Vagrant
The `provision` command used to always run the `server.yml` playbook
regardless of environment which isn't correct for development where
`dev.yml` should be used.

This sets `dev.yml` as the playbook for development and also specifies
Vagrant's generated inventory file if it exists. This is needed for
Vagrant boxes to use the proper user, host, port and private SSH key.