Run galaxy install before up cmd by default
This replaces the existing `withGalaxy` option (which defaulted to
`false`) to the inverse `noGalaxy` option that defaults to `true`.

Galaxy install was skipped by default before because Vagrant's
integration always run the full `ansible-galaxy install` even if the
roles were already installed which was annoying and slow.

But since trellis-cli has a much improved `galaxy install` command, this
runs that by default now. If you truly do want to skip this, then supply
the `--no-galaxy` option.