Add warning when project is not initialized
It's a common use case to start using trellis-cli on an existing Trellis
project (meaning the project wasn't created from `trellis new`). In
these cases, it's easy to miss that `trellis init` should be run to
create a virtualenv and install dependencies.

This adds a warning message whenever a command is run and there's no
virtualenv initialized that suggests users run `trellis init`. This is
also just a helpful reminder in case someone knows about the init
command but forgets, or the virtualenv gets deleted for some reason.

It  can be disable by the existing `TRELLIS_VENV` environment variable.