Improve arg autocomplete when flags present
Previously autocomplete would not work for positional arguments if any
flags/options were present.

Example: `trellis provision --verbose pro`

The environment argument ("prod") would not be autocompleted. This was
because of the predict functions which would act on `args.Completed`,
and unfortunately, the complete package _includes_ flags in that array.

This solution passes in the command's flag set and uses that to parse
the args. Now the predict functions acts on the positional args only,
not the full set of completed args.