Improve failure modes during new project creation
Fixes #262

This improves failure modes and error handling during the `new` command
while creating a new project. Previously it was too easy for transient
network failures to break a key part in the project creation process.
And even worse, the project would continue onto as is no failure
happened.

Now every step has proper error handling that is checked, and in the
case of a error, the process is stopped with a more helpful error
message.