key generate: use manual input for known hosts
Fixes #259

In hindsight, trellis-cli will never be able to automatically determine
the SSH known hosts without manual intervention. This approach asks the
user for input through a prompt with the default value provided by the
list of hosts parsed from Ansible. It's hopefully the best of both
worlds.

Note: the prompt library we're using doesn't allow for multi-select, so
this uses a good ol' comma-separated string for multiple hosts.