Add `droplet dns` command to create DNS records
`droplet dns` will automatically create DNS records for all the hosts
that exist on the wordpress site for an environment.

This includes multiple sites, canonical hosts, and all redirects. With a
single command you can be sure that the DNS records are up to date and
ready for SSL certificates as well.