Fix CLI config file loading
Fixes a bug where the project CLI config was only loaded if the CLI command was
run from within the `trellis` directory.

The project CLI config file was being loaded _before_ the project `Path`
was set which meant it wasn't found and loaded. Now it's loaded after the
`Path` is set.