Migrate to mholt/archives
`mholt/archiver` is now deprecated. This migrates to the `mholt/archives` replacement.

The main difference is that the application is responsible for
implementing the file handler when extracting an archive.

This also improves the `DownloadRelease` unit test which previously
didn't actually verify files within the archive were extracted on disk.