Forward Lima guest HTTP port to host
This isn't used yet but will give us more options in the future for
different "host resolvers" instead of only relying on vzNAT's networking
that makes the guest IP reachable on the host.

On Linux hosts for example, vzNAT isn't available so we would some sort
of rever proxy implementation and that can only work with the guest HTTP
port forwarded on the host.

There's a new CLI option to control this but it defaults to `true`.